+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Add Grotto/Grave entrance randomizer, for Ocarina of Time.
+- Add Grotto/Grave entrance randomizer, for both games.
 - Misc. Souls for both games, which can be shared.
 - Added a validator for Special Conditions (web-wise).